Move channel states
